+++ bot/handlers/start.py
@@ -5,7 +5,7 @@ from aiogram.utils.exceptions import ChatNotFound
 
 from bot.buttons.buttons import Buttons
 from bot.database.db import Database
-from bot.message_texts.constans import TEACHER_START_TEXT, START_TEXT, STUDENT_START_TEXT, NEW_UPDATE_TEXT
+from bot.message_texts.constans import TEACHER_START_TEXT, START_TEXT, STUDENT_START_TEXT, NEW_UPDATE_TEXT, HELP
 from bot.misc.states import MainStates
 
 
@@ -17,17 +17,20 @@ class Start:
 
 
         dp.register_message_handler(self.start_message, commands=['start'], state="*")
+        dp.register_message_handler(self.help_message, commands=['help'], state="*")
 
     async def send_upsated_bot_message(self):
-        for user_chat_id in self.db.get_chat_id_all_users():
-            user_chat_id = user_chat_id[0]
-            try:
-                 await self.bot.send_message(
-                    chat_id=user_chat_id,
-                    text=NEW_UPDATE_TEXT
-                )
-            except Exception:
-                self.db.delete_user(user_chat_id)
+        chat_ids = self.db.get_chat_id_all_users()
+        if chat_ids is not None:
+            for user_chat_id in chat_ids :
+                user_chat_id = user_chat_id[0]
+                try:
+                     await self.bot.send_message(
+                        chat_id=user_chat_id,
+                        text=NEW_UPDATE_TEXT
+                    )
+                except Exception:
+                    self.db.delete_user(user_chat_id)
 
     async def start_message(self, message : Message,state : FSMContext):
         self.db.add_user_if_not_exits(message.chat.id, message.from_user.username)
@@ -53,3 +56,8 @@ class Start:
                              reply_markup=self.buttons.get_courses_buttons())
             await state.update_data(course_id = None)
             await MainStates.registration.set()
+
+    async def help_message(self, message : Message):
+        await message.answer(
+            text=HELP
+        )
